feat: separate MainAppBar from app file
