+++ frontend/src/todos/components/AppBar.js
@@ -1,0 +1,15 @@
+import { AppBar, Toolbar, Typography } from '@mui/material'
+
+const MainAppBar = () => {
+  return (
+    <AppBar position='static' color='primary'>
+      <Toolbar>
+        <Typography variant='h6' color='inherit'>
+          Things to do
+        </Typography>
+      </Toolbar>
+    </AppBar>
+  )
+}
+
+export default MainAppBar
